requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3'
click==8.1.3; python_version >= '3.7'
cssselect==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dj-database-url==0.5.0
dj-static==0.0.6
django==3.2.12
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3; python_version >= '3'
lxml==4.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
mypy-extensions==0.4.3
pathspec==0.9.0
pillow==9.1.0
platformdirs==2.5.2; python_version >= '3.7'
psycopg2==2.9.3
pyquery==1.4.3
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
setuptools==62.1.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
static3==0.7.0
tomli==2.0.1; python_version < '3.11'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
